fix s3 file step across filesystems
When using go2chef.step.file from S3 go2chef will fail at the
os.Rename() call if /tmp is a different filesystem than the destination
of the download as golang's os.Rename calls syscall Renameat which in
turn calls link which does not support linking files across mount
points.

Instead of always creating our ioutil.TempFile() in the default
location, lets create it in the destination path instead.  This keeps
the same atomic-ness as the original implementation but ensures that the
source and destinations for os.Rename will always target the same mount
point.